lib: tool: make the staging area (and builtin defaults) a tool-only thing
