+++ src/common.h
@@ -52,18 +52,6 @@
  #include <unistd.h>
 #endif /* SJA1105_KMOD_BUILD */
 
-/* Since remapping is used internally, and many checks
- * search for a negative return code, we do that here.
- * The sign is flipped again when returning the error
- * to userspace.
- */
-#define sja1105_err_remap(old_err, new_err)           \
-	do {                                          \
-		logv("Remapping error code %d to %d", \
-		    (old_err), (new_err));            \
-		old_err = -new_err;                   \
-	} while (0);
-
 #define MAX_LINE_SIZE 2048
 
 /* Macros for conditional, error, verbose and debug logging */
